requirements.txt
flask
torch
torchvision
torchaudio
h5py
hdf5
matplotlib
matplotlib-base
opencv==4.6.0
scipy
timm==0.6.13
pillow
numpy==1.24.1
trimesh
google-cloud-storage
flask-cors
huggingface-hub
tqdm
wandb